Validate the channel on a release request
When a request is received to publish a build, a check is
now added.  All builds associated with this job are checked
to ensure that the same version of the APK has not been
published to a different channel.  If it has, an exception
is thrown and the request is rejected.