Move error console to S3
In the event of a build or publish error, copy the console
file from the Jenkins instance to S3 and return the S3 url
as the error link associated with the build.  Since S3 will
now have publish jobs in the list when publish errors occur,
changes to the S3 maintenance task had to be made to take into
consideration the new valid job names.