Add links to Jenkins Builds in the admin web
Added links to the job administration to take you to the build
and publish jobs for the job.
Added links to the index and detail views of build and release that
take you to the jenkins entry if you click on the build number

Added fields to the job table to save the location of the
jenkins base urls for the job.
Modified CopyToS3 action so that the fields are filled in for all
jobs when the scripts are rebuilt.